Sketch(): float32 to float64

+++ minhash.go
@@ -30,12 +30,12 @@ func Sketch(words []string) []byte {
 	return sketch.Bytes()
 }
 
-func Jaccard(s1, s2 []byte) float32 {
+func Jaccard(s1, s2 []byte) float64 {
 	sb1 := big.NewInt(0).SetBytes(s1)
 	sb2 := big.NewInt(0).SetBytes(s2)
 	popCnt := 0
 	for _, byte := range sb1.Xor(sb1, sb2).Bytes() {
 		popCnt += onesCount8(uint8(byte))
 	}
-	return 2.0 * (float32(k-popCnt)/float32(k) - 0.5)
+	return 2.0 * (float64(k-popCnt)/float64(k) - 0.5)
 }
